EDA and Data selection for Evaluation
I am going to use following fields of data to make my prediction regarding the severity of accidents.
1.	ADDRTYPE
2.	COLLISIONTYPE
3.	PERSONCOUNT
4.	VEHCOUNT
5.	WEATHER
6.	ROADCOND
7.	LIGHTCOND

Columns like – UNDERINFL and SPEEDING could not be used. 
Data in UNDERINFL is ambiguous and hence cannot be used while
We have lots of missing values in SPEEDING, due to which it cannot be used.
Also I will be using the data given in the Capstone project and I wont use any external data.